pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "aisurveywriter"
version = "0.0.1"
dependencies = [
    "selenium",
    "pandas",
    "numpy",
    "undetected-chromedriver",
    "fake-useragent",
    "pynvim",
    "setuptools",
    "langchain",
    "langchain-google-genai",
    "langchain-openai",
    "langchain-ollama",
    "langchain-community",
    "faiss-cpu",
    "pypdf",
    "bibtexparser",
    "gradio",
]
requires-python = ">=3.11"
description = "LLM tool to write Survey Papers based on PDF references"